Stubbed subs no longer have CvOUTSIDE set.
t/code.t was testing whether Devel::Size was following CvOUTSIDE by relying
on the implementation detail that CvOUTSIDE is set for a subroutine that is
explicitly stubbed, and not set for one that is implicitly stubbed at runtime.

v5.17.0 changed the compile-time behaviour so that no stubbed subroutine has
CvOUTSIDE set, which breaks the assumption of this test. Hence change the test
to verify the new behaviour. However, this leaves us currently with no test
for CvOUTSIDE.

(Based on a patch by Zefram. see RT #83903 and perl's commit 8be227ab5eaa23f2)