A more robust test for the size of subroutines in packages which import.
A test in t/recurse.t broke because Test::More::is() had been refactored.
Obviously that wasn't correct. Ultimately the problem was because the test
as-was wasn't minimally testing the thing that it was intended to test. It
should be now. Comments in the test explain the problem.

Reported as CPAN #102910.