Try to avoid strange failures when 1 == 1 doesn't give us PL_sv_yes.
The intent of these tests is to confirm that the interpreter variables
PL_sv_undef, PL_sv_no and PL_sv_yes are not reported as part of the size.
It seems that on some older threaded perls with newer versions of
Test::More, we can end up with 1 == 1 not passing PL_sv_yes into
total_size(), presuambly because some code in Test::More has also caused
PL_sv_yes to end up in a pad slot.

The ugly (and possibly fragile) fix seems to be to ensure that our test code
is the first thing to get compiled, by moving the load of Test::More later.